requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
backports.ssl-match-hostname==3.5.0.1  # via recurly
contextlib2==0.5.5        # via vcrpy
funcsigs==1.0.2           # via mock
iso8601==0.1.12           # via recurly
mock==2.0.0               # via vcrpy
pbr==3.1.1                # via mock
pyyaml==3.12              # via vcrpy
recurly==2.7.0
six==1.11.0               # via mock, recurly, vcrpy
vcrpy==1.11.1
wrapt==1.10.11            # via vcrpy
